R-objects related to metabolomics data on patients with Alzheimer's Disease

ADdata contains 3 objects related to metabolomics data on patients with Alzheimer's Disease (AD).

Source

de Leeuw, F., Peeters, C.F.W., Kester, M.I., Harms, A.C., Struys, E., Hankemeijer, T., van Vlijmen, H.W.T., van Duijn, C.M., Scheltens, P., Demirkan, A., van de Wiel, M.A., van der Flier, W.M., and Teunissen, C.E. (2017). Blood-based metabolic signatures in Alzheimer's Disease. Alzheimer's & Dementia: Diagnosis, Assessment & Disease Monitoring, 8: 196-207.

Details

ADmetabolites is a matrix containing metabolic expressions of 230 metabolites (rows) on 127 samples (columns).

sampleInfo is a data.frame containing information on the samples. Information pertains to diagnosis: AD class 1 or AD class 2.

variableInfo is a data.frame containing information on the metabolic features. Information pertains to compound families: Amines, organic acids, lipids, and oxidative stress compounds.

See description.

Author

Carel F.W. Peeters <cf.peeters@vumc.nl>

Examples

data(ADdata)

## Look at sample information
sampleInfo

#>              Diagnosis ApoEClass
#> Sample_1   Probable AD   Class 1
#> Sample_2   Probable AD   Class 1
#> Sample_3   Probable AD   Class 2
#> Sample_4   Probable AD   Class 2
#> Sample_5   Probable AD   Class 2
#> Sample_6   Probable AD   Class 1
#> Sample_7   Probable AD   Class 2
#> Sample_8   Probable AD   Class 2
#> Sample_9   Probable AD   Class 2
#> Sample_10  Probable AD   Class 2
#> Sample_11  Probable AD   Class 1
#> Sample_12  Probable AD   Class 2
#> Sample_13  Probable AD   Class 2
#> Sample_14  Probable AD   Class 2
#> Sample_15  Probable AD   Class 1
#> Sample_16  Probable AD   Class 2
#> Sample_17  Probable AD   Class 2
#> Sample_18  Probable AD   Class 1
#> Sample_19  Probable AD   Class 1
#> Sample_20  Probable AD   Class 2
#> Sample_21  Probable AD   Class 2
#> Sample_22  Probable AD   Class 2
#> Sample_23  Probable AD   Class 1
#> Sample_24  Probable AD   Class 2
#> Sample_25  Probable AD   Class 2
#> Sample_26  Probable AD   Class 1
#> Sample_27  Probable AD   Class 1
#> Sample_28  Probable AD   Class 1
#> Sample_29  Probable AD   Class 1
#> Sample_30  Probable AD   Class 1
#> Sample_31  Probable AD   Class 2
#> Sample_32  Probable AD   Class 2
#> Sample_33  Probable AD   Class 1
#> Sample_34  Probable AD   Class 2
#> Sample_35  Probable AD   Class 2
#> Sample_36  Probable AD   Class 2
#> Sample_37  Probable AD   Class 2
#> Sample_38  Probable AD   Class 1
#> Sample_39  Probable AD   Class 1
#> Sample_40  Probable AD   Class 1
#> Sample_41  Probable AD   Class 2
#> Sample_42  Probable AD   Class 2
#> Sample_43  Probable AD   Class 2
#> Sample_44  Probable AD   Class 2
#> Sample_45  Probable AD   Class 2
#> Sample_46  Probable AD   Class 1
#> Sample_47  Probable AD   Class 1
#> Sample_48  Probable AD   Class 1
#> Sample_49  Probable AD   Class 2
#> Sample_50  Probable AD   Class 1
#> Sample_51  Probable AD   Class 2
#> Sample_52  Probable AD   Class 2
#> Sample_53  Probable AD   Class 1
#> Sample_54  Probable AD   Class 2
#> Sample_55  Probable AD   Class 2
#> Sample_56  Probable AD   Class 1
#> Sample_57  Probable AD   Class 2
#> Sample_58  Probable AD   Class 2
#> Sample_59  Probable AD   Class 1
#> Sample_60  Probable AD   Class 2
#> Sample_61  Probable AD   Class 2
#> Sample_62  Probable AD   Class 2
#> Sample_63  Probable AD   Class 2
#> Sample_64  Probable AD   Class 2
#> Sample_65  Probable AD   Class 2
#> Sample_66  Probable AD   Class 2
#> Sample_67  Probable AD   Class 2
#> Sample_68  Probable AD   Class 1
#> Sample_69  Probable AD   Class 2
#> Sample_70  Probable AD   Class 1
#> Sample_71  Probable AD   Class 2
#> Sample_72  Probable AD   Class 2
#> Sample_73  Probable AD   Class 2
#> Sample_74  Probable AD   Class 1
#> Sample_75  Probable AD   Class 2
#> Sample_76  Probable AD   Class 2
#> Sample_77  Probable AD   Class 2
#> Sample_78  Probable AD   Class 2
#> Sample_79  Probable AD   Class 2
#> Sample_80  Probable AD   Class 2
#> Sample_81  Probable AD   Class 2
#> Sample_82  Probable AD   Class 1
#> Sample_83  Probable AD   Class 1
#> Sample_84  Probable AD   Class 2
#> Sample_85  Probable AD   Class 1
#> Sample_86  Probable AD   Class 2
#> Sample_87  Probable AD   Class 1
#> Sample_88  Probable AD   Class 1
#> Sample_89  Probable AD   Class 2
#> Sample_90  Probable AD   Class 2
#> Sample_91  Probable AD   Class 2
#> Sample_92  Probable AD   Class 2
#> Sample_93  Probable AD   Class 2
#> Sample_94  Probable AD   Class 2
#> Sample_95  Probable AD   Class 2
#> Sample_96  Probable AD   Class 2
#> Sample_97  Probable AD   Class 2
#> Sample_98  Probable AD   Class 2
#> Sample_99  Probable AD   Class 2
#> Sample_100 Probable AD   Class 2
#> Sample_101 Probable AD   Class 2
#> Sample_102 Probable AD   Class 2
#> Sample_103 Probable AD   Class 2
#> Sample_104 Probable AD   Class 1
#> Sample_105 Probable AD   Class 2
#> Sample_106 Probable AD   Class 2
#> Sample_107 Probable AD   Class 2
#> Sample_108 Probable AD   Class 2
#> Sample_109 Probable AD   Class 2
#> Sample_110 Probable AD   Class 2
#> Sample_111 Probable AD   Class 1
#> Sample_112 Probable AD   Class 1
#> Sample_113 Probable AD   Class 2
#> Sample_114 Probable AD   Class 1
#> Sample_115 Probable AD   Class 2
#> Sample_116 Probable AD   Class 2
#> Sample_117 Probable AD   Class 2
#> Sample_118 Probable AD   Class 2
#> Sample_119 Probable AD   Class 1
#> Sample_120 Probable AD   Class 2
#> Sample_121 Probable AD   Class 1
#> Sample_122 Probable AD   Class 2
#> Sample_123 Probable AD   Class 2
#> Sample_124 Probable AD   Class 1
#> Sample_125 Probable AD   Class 2
#> Sample_126 Probable AD   Class 1
#> Sample_127 Probable AD   Class 2

## Look at feature information
variableInfo

#>                        Family FamilyCode
#> Amine_1                Amines          1
#> Amine_2                Amines          1
#> Amine_3                Amines          1
#> Amine_4                Amines          1
#> Amine_5                Amines          1
#> Amine_6                Amines          1
#> Amine_7                Amines          1
#> Amine_8                Amines          1
#> Amine_9                Amines          1
#> Amine_10               Amines          1
#> Amine_11               Amines          1
#> Amine_12               Amines          1
#> Amine_13               Amines          1
#> Amine_14               Amines          1
#> Amine_15               Amines          1
#> Amine_16               Amines          1
#> Amine_17               Amines          1
#> Amine_18               Amines          1
#> Amine_19               Amines          1
#> Amine_20               Amines          1
#> Amine_21               Amines          1
#> Amine_22               Amines          1
#> Amine_23               Amines          1
#> Amine_24               Amines          1
#> Amine_25               Amines          1
#> Amine_26               Amines          1
#> Amine_27               Amines          1
#> Amine_28               Amines          1
#> Amine_29               Amines          1
#> Amine_30               Amines          1
#> Amine_31               Amines          1
#> Amine_32               Amines          1
#> Amine_33               Amines          1
#> Amine_34               Amines          1
#> Amine_35               Amines          1
#> Amine_36               Amines          1
#> Amine_37               Amines          1
#> Amine_38               Amines          1
#> Amine_39               Amines          1
#> Amine_40               Amines          1
#> Amine_41               Amines          1
#> Amine_42               Amines          1
#> Amine_43               Amines          1
#> Amine_44               Amines          1
#> Amine_45               Amines          1
#> Amine_46               Amines          1
#> Amine_47               Amines          1
#> Amine_48               Amines          1
#> Amine_49               Amines          1
#> Amine_50               Amines          1
#> Amine_51               Amines          1
#> Amine_52               Amines          1
#> Amine_53               Amines          1
#> Org.Acid_1      Organic Acids          2
#> Org.Acid_2      Organic Acids          2
#> Org.Acid_3      Organic Acids          2
#> Org.Acid_4      Organic Acids          2
#> Org.Acid_5      Organic Acids          2
#> Org.Acid_6      Organic Acids          2
#> Org.Acid_7      Organic Acids          2
#> Org.Acid_8      Organic Acids          2
#> Org.Acid_9      Organic Acids          2
#> Org.Acid_10     Organic Acids          2
#> Org.Acid_11     Organic Acids          2
#> Org.Acid_12     Organic Acids          2
#> Org.Acid_13     Organic Acids          2
#> Org.Acid_14     Organic Acids          2
#> Org.Acid_15     Organic Acids          2
#> Org.Acid_16     Organic Acids          2
#> Org.Acid_17     Organic Acids          2
#> Org.Acid_18     Organic Acids          2
#> Org.Acid_19     Organic Acids          2
#> Org.Acid_20     Organic Acids          2
#> Org.Acid_21     Organic Acids          2
#> Org.Acid_22     Organic Acids          2
#> Lip_1                  Lipids          3
#> Lip_2                  Lipids          3
#> Lip_3                  Lipids          3
#> Lip_4                  Lipids          3
#> Lip_5                  Lipids          3
#> Lip_6                  Lipids          3
#> Lip_7                  Lipids          3
#> Lip_8                  Lipids          3
#> Lip_9                  Lipids          3
#> Lip_10                 Lipids          3
#> Lip_11                 Lipids          3
#> Lip_12                 Lipids          3
#> Lip_13                 Lipids          3
#> Lip_14                 Lipids          3
#> Lip_15                 Lipids          3
#> Lip_16                 Lipids          3
#> Lip_17                 Lipids          3
#> Lip_18                 Lipids          3
#> Lip_19                 Lipids          3
#> Lip_20                 Lipids          3
#> Lip_21                 Lipids          3
#> Lip_22                 Lipids          3
#> Lip_23                 Lipids          3
#> Lip_24                 Lipids          3
#> Lip_25                 Lipids          3
#> Lip_26                 Lipids          3
#> Lip_27                 Lipids          3
#> Lip_28                 Lipids          3
#> Lip_29                 Lipids          3
#> Lip_30                 Lipids          3
#> Lip_31                 Lipids          3
#> Lip_32                 Lipids          3
#> Lip_33                 Lipids          3
#> Lip_34                 Lipids          3
#> Lip_35                 Lipids          3
#> Lip_36                 Lipids          3
#> Lip_37                 Lipids          3
#> Lip_38                 Lipids          3
#> Lip_39                 Lipids          3
#> Lip_40                 Lipids          3
#> Lip_41                 Lipids          3
#> Lip_42                 Lipids          3
#> Lip_43                 Lipids          3
#> Lip_44                 Lipids          3
#> Lip_45                 Lipids          3
#> Lip_46                 Lipids          3
#> Lip_47                 Lipids          3
#> Lip_48                 Lipids          3
#> Lip_49                 Lipids          3
#> Lip_50                 Lipids          3
#> Lip_51                 Lipids          3
#> Lip_52                 Lipids          3
#> Lip_53                 Lipids          3
#> Lip_54                 Lipids          3
#> Lip_55                 Lipids          3
#> Lip_56                 Lipids          3
#> Lip_57                 Lipids          3
#> Lip_58                 Lipids          3
#> Lip_59                 Lipids          3
#> Lip_60                 Lipids          3
#> Lip_61                 Lipids          3
#> Lip_62                 Lipids          3
#> Lip_63                 Lipids          3
#> Lip_64                 Lipids          3
#> Lip_65                 Lipids          3
#> Lip_66                 Lipids          3
#> Lip_67                 Lipids          3
#> Lip_68                 Lipids          3
#> Lip_69                 Lipids          3
#> Lip_70                 Lipids          3
#> Lip_71                 Lipids          3
#> Lip_72                 Lipids          3
#> Lip_73                 Lipids          3
#> Lip_74                 Lipids          3
#> Lip_75                 Lipids          3
#> Lip_76                 Lipids          3
#> Lip_77                 Lipids          3
#> Lip_78                 Lipids          3
#> Lip_79                 Lipids          3
#> Lip_80                 Lipids          3
#> Lip_81                 Lipids          3
#> Lip_82                 Lipids          3
#> Lip_83                 Lipids          3
#> Lip_84                 Lipids          3
#> Lip_85                 Lipids          3
#> Lip_86                 Lipids          3
#> Lip_87                 Lipids          3
#> Lip_88                 Lipids          3
#> Lip_89                 Lipids          3
#> Lip_90                 Lipids          3
#> Lip_91                 Lipids          3
#> Lip_92                 Lipids          3
#> Lip_93                 Lipids          3
#> Lip_94                 Lipids          3
#> Lip_95                 Lipids          3
#> Lip_96                 Lipids          3
#> Lip_97                 Lipids          3
#> Lip_98                 Lipids          3
#> Lip_99                 Lipids          3
#> Lip_100                Lipids          3
#> Lip_101                Lipids          3
#> Lip_102                Lipids          3
#> Lip_103                Lipids          3
#> Lip_104                Lipids          3
#> Lip_105                Lipids          3
#> Lip_106                Lipids          3
#> Lip_107                Lipids          3
#> Lip_108                Lipids          3
#> Lip_109                Lipids          3
#> Lip_110                Lipids          3
#> Lip_111                Lipids          3
#> Lip_112                Lipids          3
#> Lip_113                Lipids          3
#> Lip_114                Lipids          3
#> Lip_115                Lipids          3
#> Lip_116                Lipids          3
#> Ox.Stress_1  Oxidative Stress          4
#> Ox.Stress_2  Oxidative Stress          4
#> Ox.Stress_3  Oxidative Stress          4
#> Ox.Stress_4  Oxidative Stress          4
#> Ox.Stress_5  Oxidative Stress          4
#> Ox.Stress_6  Oxidative Stress          4
#> Ox.Stress_7  Oxidative Stress          4
#> Ox.Stress_8  Oxidative Stress          4
#> Ox.Stress_9  Oxidative Stress          4
#> Ox.Stress_10 Oxidative Stress          4
#> Ox.Stress_11 Oxidative Stress          4
#> Ox.Stress_12 Oxidative Stress          4
#> Ox.Stress_13 Oxidative Stress          4
#> Ox.Stress_14 Oxidative Stress          4
#> Ox.Stress_15 Oxidative Stress          4
#> Ox.Stress_16 Oxidative Stress          4
#> Ox.Stress_17 Oxidative Stress          4
#> Ox.Stress_18 Oxidative Stress          4
#> Ox.Stress_19 Oxidative Stress          4
#> Ox.Stress_20 Oxidative Stress          4
#> Ox.Stress_21 Oxidative Stress          4
#> Ox.Stress_22 Oxidative Stress          4
#> Ox.Stress_23 Oxidative Stress          4
#> Ox.Stress_24 Oxidative Stress          4
#> Ox.Stress_25 Oxidative Stress          4
#> Ox.Stress_26 Oxidative Stress          4
#> Ox.Stress_27 Oxidative Stress          4
#> Ox.Stress_28 Oxidative Stress          4
#> Ox.Stress_29 Oxidative Stress          4
#> Ox.Stress_30 Oxidative Stress          4
#> Ox.Stress_31 Oxidative Stress          4
#> Ox.Stress_32 Oxidative Stress          4
#> Ox.Stress_33 Oxidative Stress          4
#> Ox.Stress_34 Oxidative Stress          4
#> Ox.Stress_35 Oxidative Stress          4
#> Ox.Stress_36 Oxidative Stress          4
#> Ox.Stress_37 Oxidative Stress          4
#> Ox.Stress_38 Oxidative Stress          4
#> Ox.Stress_39 Oxidative Stress          4