Generate a (data-driven) default target for usage in ridge-type shrinkage estimation

Function that generates a (data-driven) default target for usage in (type I) ridge shrinkage estimation of the precision matrix (see ridgeP). The target that is generated is to be understood in precision terms. Most options for target generation result in a target that implies a situation of rotation equivariant estimation (see ridgeP).

default.target(S, type = "DAIE", fraction = 1e-04, const)

Arguments

S
Sample covariance matrix.

type
A character determining the type of default target. Must be one of: "DAIE", "DIAES", "DUPV", "DAPV", "DCPV", "DEPV", "Null".

fraction
A numeric indicating the fraction of the largest eigenvalue below which an eigenvalue is considered zero.

const
A numeric constant representing the partial variance.

Value

Function returns a target matrix.

Details

The function can generate the following default target matrices:

DAIE: Diagonal matrix with average of inverse nonzero eigenvalues of S as entries;

DIAES: Diagonal matrix with inverse of average of eigenvalues of S as entries;

DUPV: Diagonal matrix with unit partial variance as entries (identity matrix);

DAPV: Diagonal matrix with average of inverse variances of S as entries;

DCPV: Diagonal matrix with constant partial variance as entries. Allows one to use other constant than DAIE, DIAES, DUPV, DAPV, and in a sense Null;

DEPV: Diagonal matrix with the inverse variances of S as entries;

Null: Null matrix.

The targets DUPV, DCPV, and Null are not data-driven in the sense that the input matrix S only provides information on the size of the desired target. The targets DAIE, DIAES, DAPV, and DEPV are data-driven in the sense that the input matrix S provides the information for the diagonal entries. The argument fraction is only used when type = "DAIE". The argument const is only used when type = "DCPV". All types except DEPV and Null lead to rotation equivariant alternative and archetypal Type I ridge estimators. The target Null also leads to a rotation equivariant alternative Type II ridge estimator (see ridgeP). Note that the DIAES, DAPV, and DEPV targets amount to the identity matrix when the sample covariance matrix S is standardized to be the correlation matrix. The same goes, naturally, for the DCPV target when const is specified to be 1.

References

van Wieringen, W.N. & Peeters, C.F.W. (2016). Ridge Estimation of Inverse Covariance Matrices from High-Dimensional Data, Computational Statistics & Data Analysis, vol. 103: 284-303. Also available as arXiv:1403.0904v3 [stat.ME].

See also

Author

Carel F.W. Peeters <cf.peeters@vumc.nl>, Wessel N. van Wieringen

Examples

## Obtain some (high-dimensional) data
p = 25
n = 10
set.seed(333)
X = matrix(rnorm(n*p), nrow = n, ncol = p)
colnames(X)[1:25] = letters[1:25]
Cx <- covML(X)

## Obtain default diagonal target matrix
default.target(Cx)

#>           a         b         c         d         e         f         g
#> a 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> b 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> c 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000
#> d 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000
#> e 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000
#> f 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000
#> g 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251
#> h 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> i 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> j 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> k 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> l 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> m 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> n 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> o 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> p 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> q 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> r 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> s 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> t 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> u 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> v 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> w 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> x 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> y 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           h         i         j         k         l         m         n
#> a 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> b 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> c 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> d 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> e 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> f 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> g 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> h 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> i 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> j 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000
#> k 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000
#> l 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000
#> m 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000
#> n 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251
#> o 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> p 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> q 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> r 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> s 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> t 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> u 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> v 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> w 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> x 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> y 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           o         p         q         r         s         t         u
#> a 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> b 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> c 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> d 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> e 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> f 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> g 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> h 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> i 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> j 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> k 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> l 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> m 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> n 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> o 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> p 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> q 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000 0.0000000
#> r 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000 0.0000000
#> s 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000 0.0000000
#> t 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251 0.0000000
#> u 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.7495251
#> v 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> w 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> x 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#> y 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000 0.0000000
#>           v         w         x         y
#> a 0.0000000 0.0000000 0.0000000 0.0000000
#> b 0.0000000 0.0000000 0.0000000 0.0000000
#> c 0.0000000 0.0000000 0.0000000 0.0000000
#> d 0.0000000 0.0000000 0.0000000 0.0000000
#> e 0.0000000 0.0000000 0.0000000 0.0000000
#> f 0.0000000 0.0000000 0.0000000 0.0000000
#> g 0.0000000 0.0000000 0.0000000 0.0000000
#> h 0.0000000 0.0000000 0.0000000 0.0000000
#> i 0.0000000 0.0000000 0.0000000 0.0000000
#> j 0.0000000 0.0000000 0.0000000 0.0000000
#> k 0.0000000 0.0000000 0.0000000 0.0000000
#> l 0.0000000 0.0000000 0.0000000 0.0000000
#> m 0.0000000 0.0000000 0.0000000 0.0000000
#> n 0.0000000 0.0000000 0.0000000 0.0000000
#> o 0.0000000 0.0000000 0.0000000 0.0000000
#> p 0.0000000 0.0000000 0.0000000 0.0000000
#> q 0.0000000 0.0000000 0.0000000 0.0000000
#> r 0.0000000 0.0000000 0.0000000 0.0000000
#> s 0.0000000 0.0000000 0.0000000 0.0000000
#> t 0.0000000 0.0000000 0.0000000 0.0000000
#> u 0.0000000 0.0000000 0.0000000 0.0000000
#> v 0.7495251 0.0000000 0.0000000 0.0000000
#> w 0.0000000 0.7495251 0.0000000 0.0000000
#> x 0.0000000 0.0000000 0.7495251 0.0000000
#> y 0.0000000 0.0000000 0.0000000 0.7495251