Print and summarize fusion test

Print and summary functions for the fusion test performed by fused.test.

# S3 method for ptest
print(x, digits = 4L, ...)

# S3 method for ptest
summary(object, ...)

Arguments

x, object
The object to print or summarize. Usually the output of fused.test.

digits
An integer controlling the number of printed digits.

...
Arguments passed on. In summary.ptest the arguments are passed to print.ptest. In print.ptest are passed to the standard summary function.

Value

Invisibly returns the object.

References

Bilgrau, A.E., Peeters, C.F.W., Eriksen, P.S., Boegsted, M., and van Wieringen, W.N. (2020). Targeted Fused Ridge Estimation of Inverse Covariance Matrices from Multiple High-Dimensional Data Classes. Journal of Machine Learning Research, 21(26): 1-52.

See also

Author

Anders Ellern Bilgrau, Carel F.W. Peeters <carel.peeters@wur.nl>, Wessel N. van Wieringen

Examples

ns <- c(10, 5, 23)
Ylist <- createS(ns, p = 15, topology = "banded", dataset = TRUE)

# Use the identity target matrix for each class
Tlist <- replicate(length(ns), diag(15), simplify = FALSE)

# Do the test
lam <- matrix(10, 3, 3)
diag(lam) <- 1
ft <- fused.test(Ylist, Tlist, lambda = lam, n.permutations = 500)