Ridge estimation for high-dimensional precision matrices

Package contains proper L2-penalized ML estimators for the precision matrix as well as supporting functions to employ these estimators in a (integrative or meta-analytic) graphical modeling setting.

Details

The main function of the package is ridgeP which enables archetypal and proper alternative ML ridge estimation of the precision matrix. The alternative ridge estimators can be found in van Wieringen and Peeters (2015) and encapsulate both target and non-target shrinkage for the multivariate normal precision matrix. The estimators are analytic and enable estimation in large \(p\) small \(n\) settings. Supporting functions to employ these estimators in a graphical modeling setting are also given. These supporting functions enable, a.o., the determination of the optimal value of the penalty parameter, the determination of the support of a shrunken precision estimate, as well as various visualization options.

The package has a modular setup. The core module (rags2ridges.R) contains the functionality stated above. The fused module (rags2ridgesFused.R) extends the functionality of the core module to the joint estimation of multiple precision matrices from (aggregated) high-dimensional data consisting of distinct classes. The result is a targeted fused ridge estimator that is of use when the precision matrices of the constituent classes are believed to chiefly share the same structure while potentially differing in a number of locations of interest. The fused module also contains supporting functions for integrative or meta-analytic Gaussian graphical modeling. The third module is the miscellaneous module (rags2RidgesMisc.R) which contains assorted hidden functions.

Function overview core module:

Function for (proper) ridge estimation of the precision matrix

ridgeP

Functions for penalty parameter selection

CNplot

optPenalty.aLOOCV

optPenalty.kCV

optPenalty.kCVauto

Functions for loss/entropy/fit evaluation

evaluateSfit

KLdiv

loss

Functions for block-independence testing

GGMblockNullPenalty

GGMblockTest

Function for support determination

sparsify

Functions for (network) visualization

edgeHeat

ridgePathS

Ugraph

Functions for topology statistics

GGMmutualInfo

GGMnetworkStats

GGMpathStats

Wrapper function

fullMontyS

Support functions

adjacentMat

covML

covMLknown

default.target

evaluateS

pcor

symm

Function overview fused module:

Function for targeted fused ridge estimation of multiple precision matrices

ridgeP.fused

Function for fused penalty parameter selection

optPenalty.fused

Functions for loss/entropy/fit evaluation

KLdiv.fused

NLL

Function for testing the necessity of fusion

fused.test

Function for support determination

sparsify.fused

Functions for topology statistics

GGMnetworkStats.fused

GGMpathStats.fused

Support functions

createS

default.penalty

default.target.fused

getKEGGPathway

isSymmetricPD

is.Xlist

kegg.target

plot.ptest

pooledS

print.optPenaltyFusedGrid

print.ptest

rmvnormal

Calls of interest to miscellaneous module:

rags2ridges:::.TwoCents() ~~(Unsolicited advice)

rags2ridges:::.Brooke() ~~(Endorsement)

rags2ridges:::.JayZScore() ~~(The truth)

rags2ridges:::.theHoff() ~~(Wish)

rags2ridges:::.rags2logo() ~~(Warm welcome)

References

Peeters, C.F.W., Bilgrau, A.E., and van Wieringen, W.N. (2022). rags2ridges: A One-Stop-l2-Shop for Graphical Modeling of High-Dimensional Precision Matrices. Journal of Statistical Software, vol. 102(4): 1-32.

Bilgrau, A.E., Peeters, C.F.W., Eriksen, P.S., Boegsted, M., and van Wieringen, W.N. (2020). Targeted Fused Ridge Estimation of Inverse Covariance Matrices from Multiple High-Dimensional Data Classes. Journal of Machine Learning Research, 21(26): 1-52. Also available as arXiv:1509.07982v2 [stat.ME].

Peeters, C.F.W., van de Wiel, M.A., & van Wieringen, W.N. (2020). The Spectral Condition Number Plot for Regularization Parameter Evaluation. Computational Statistics, 35: 629-646. Also available as arXiv:1608.04123 [stat.CO].

van Wieringen, W.N. & Peeters, C.F.W. (2016). Ridge Estimation of Inverse Covariance Matrices from High-Dimensional Data. Computational Statistics & Data Analysis, vol. 103: 284-303. Also available as arXiv:1403.0904v3 [stat.ME].

van Wieringen, W.N. & Peeters, C.F.W. (2015). Application of a New Ridge Estimator of the Inverse Covariance Matrix to the Reconstruction of Gene-Gene Interaction Networks. In: di Serio, C., Lio, P., Nonis, A., and Tagliaferri, R. (Eds.) `Computational Intelligence Methods for Bioinformatics and Biostatistics'. Lecture Notes in Computer Science, vol. 8623. Springer, pp. 170-179.

Author

Carel F.W. Peeters, Anders Ellern Bilgrau, Wessel, N. van Wieringen
Maintainer: Carel F.W. Peeters <carel.peeters@wur.nl>