Download KEGG pathway

Download information and graph object of a given pathway from the Kyoto Encyclopedia of Genes and Genomes (KEGG) database.

getKEGGPathway(kegg.id)

Arguments

kegg.id
A character giving the KEGG ID, e.g. "map04210", "map04064", "map04115". Can be prefixed with "path:".

Value

Returns a list with entries:

df

A data.frame description of the KEGG pathway.

graph

The KEGG pathway represented as a graphNEL object.

Details

Usage of this function requires an internet connection. The igraph objects can be obtained with igraph::igraph.from.graphNEL. The moral graph can be obtained with gRbase::moralize. To obtain the adjacency matrix, use gRbase::as.adjMAT or igraph::get.adjacency

Note

It is currently necessary to require("KEGGgraph") (or require("KEGGgraph")) due to a bug in KEGGgraph.

References

https://www.genome.jp/kegg/

See also

Author

Anders Ellern Bilgrau, Carel F.W. Peeters <carel.peeters@wur.nl>, Wessel N. van Wieringen

Examples

if (FALSE) {
if (require("KEGGgraph")) {
  getKEGGPathway("map04064")
}
}