Test if fused list-formats are correctly used

Function to check if the argument submits to the various list-formats used by the fused ridge estimator and related functions are correct. That is, it tests if generic fused list arguments (such as Slist, Tlist, Plist, Ylist) are properly formatted.

is.Xlist(Xlist, Ylist = FALSE, semi = FALSE)

Arguments

Xlist

A list of precision matrices of equal size (Plist), sample covariance matrices (Slist), data matrices (Ylist)

Ylist

logical. Is Xlist a list of data matrices with the same number of columns (Ylist).

semi

logical. Should the matrices in the list be tested to be positive semi definite or positive definite?

Value

Returns TRUE if all tests are passed, throws error if not.

References

Bilgrau, A.E., Peeters, C.F.W., Eriksen, P.S., Boegsted, M., and van Wieringen, W.N. (2020). Targeted Fused Ridge Estimation of Inverse Covariance Matrices from Multiple High-Dimensional Data Classes. Journal of Machine Learning Research, 21(26): 1-52.

van Wieringen, W.N. & Peeters, C.F.W. (2016). Ridge Estimation of Inverse Covariance Matrices from High-Dimensional Data, Computational Statistics & Data Analysis, vol. 103: 284-303. Also available as arXiv:1403.0904v3 [stat.ME].

See also

Author

Anders Bilgrau, Carel F.W. Peeters <carel.peeters@wur.nl>, Wessel N. van Wieringen

Examples

Slist <- createS(n = c(4, 6, 9), p = 10)
is.Xlist(Slist, semi = TRUE)
#> [1] TRUE