Construct target matrix from KEGG

Construct a target matrix by combining topology information from the Kyoto Encyclopedia of Genes and Genomes (KEGG) database and pilot data.

kegg.target(
  Y,
  kegg.id,
  method = "linreg",
  organism = "hsa",
  graph = getKEGGPathway(kegg.id)$graph
)

Arguments

Y
The complete observation matrix of observations with variables in columns. The column names should be on the form e.g. "hsa:3988" ("<organism>:<Entrez id>"). It can however also be just the Entrez id with or without the post-fixed "_at" and then the specified organism will be assumed.

kegg.id
A character giving the KEGG ID, e.g. "map04210", "map04064", or "map04115".

method
The method for estimating the non-zero entries moralized graph of the KEGG topology. Currently, only "linreg" is implemented.

organism
A character giving the organism, the default is "hsa" (homo-sapiens).

graph
A graphNEL object specifying the topology of the pathway. Can be used to avoid repeatedly downloading the information.

Value

Returns a target matrix with size depending on the kegg.id.

Details

The function estimates the precision matrix based on the topology given by the KEGG database. Requires a connection to the internet.

Note

It is currently nessesary to require("KEGGgraph") (or require("KEGGgraph")) due to a bug in KEGGgraph.

References

https://www.genome.jp/kegg/

See also

Author

Anders Ellern Bilgrau, Carel F.W. Peeters <cf.peeters@vumc.nl>, Wessel N. van Wieringen

Examples

if (FALSE) {
if (require("KEGGgraph")) {
kegg.g <- getKEGGPathway("map04115")$graph

# Create some toy data with the correct names
Y <- createS(n = 10, p = numNodes(kegg.g), dataset = TRUE)
colnames(Y) <- nodes(kegg.g)

T <- kegg.target(Y, "map04115")
print(T[1:10, 1:10])
}
}