Moments of the sample covariance matrix.

Calculates the moments of the sample covariance matrix. It assumes that the summands (the outer products of the samples' random data vector) that constitute the sample covariance matrix follow a Wishart-distribution with scale parameter \(\mathbf{\Sigma}\) and shape parameter \(\nu\). The latter is equal to the number of summands in the sample covariance estimate.

momentS(Sigma, shape, moment = 1)

Arguments

Sigma

Positive-definite matrix, the scale parameter \(\mathbf{\Sigma}\) of the Wishart distribution.

shape

A numeric, the shape parameter \(\nu\) of the Wishart distribution. Should exceed the number of variates (number of rows or columns of Sigma).

moment

An integer. Should be in the set \(\{-4, -3, -2, -1, 0, 1, 2, 3, 4\}\) (only those are explicitly specified in Lesac, Massam, 2004).

Value

The \(r\)-th moment of a sample covariance matrix: \(E(\mathbf{S}^r)\).

References

Lesac, G., Massam, H. (2004), "All invariant moments of the Wishart distribution", Scandinavian Journal of Statistics, 31(2), 295-318.

Author

Wessel N. van Wieringen.

Examples

# create scale parameter
Sigma <- matrix(c(1, 0.5, 0, 0.5, 1, 0, 0, 0, 1), byrow=TRUE, ncol=3)

# evaluate expectation of the square of a sample covariance matrix
# that is assumed to Wishart-distributed random variable with the
# above scale parameter Sigma and shape parameter equal to 40.
momentS(Sigma, 40, 2)
#>         [,1]    [,2] [,3]
#> [1,] 1.35625 1.06250  0.0
#> [2,] 1.06250 1.35625  0.0
#> [3,] 0.00000 0.00000  1.1